+++ inst/d3js/generatePlots.js
@@ -27,7 +27,7 @@ if (svg.select("#SV").empty()) {
 } else {
   SV = svg.select("#SV");
 }
-mapIdPlotFunction.SV = shapValues;
+mapIdPlotFunction.SV = shapleyValues;
 
 if (svg.select("#CP").empty()) {
   CP = svg.append("g")
@@ -57,7 +57,7 @@ if (svg.select("#PD").empty()) {
 } else {
   PD = svg.select("#PD");
 }
-mapIdPlotFunction.PD = partialDependency;
+mapIdPlotFunction.PD = partialDependence;
 
 if (svg.select("#AD").empty()) {
   AD = svg.append("g")
@@ -67,7 +67,7 @@ if (svg.select("#AD").empty()) {
 } else {
   AD = svg.select("#AD");
 }
-mapIdPlotFunction.AD = accumulatedDependency;
+mapIdPlotFunction.AD = accumulatedDependence;
 
 if (svg.select("#FD").empty()) {
   FD = svg.append("g")
@@ -282,7 +282,7 @@ function breakDown() {
             switch (d.variable) {
               case "intercept":
               case "prediction":
-                return d.cummulative;
+                return d.cumulative;
               default:
                 return d.sign === "-1" ? d.contribution : "+"+d.contribution;
             }
@@ -338,7 +338,7 @@ function breakDown() {
              .on('mouseout', tooltip.hide);
 }
 
-function shapValues() {
+function shapleyValues() {
 
   let tObservationId = CLICKED_OBSERVATION_ID,
       tData = obsData[tObservationId];
@@ -478,15 +478,15 @@ function shapValues() {
 
   ctbLabel.append("text")
           .attr("class", "axisLabel")
-          .attr("x", d => d.contribution < 0
+          .attr("x", d => d.contribution > 0
                           ? x(d.barStart) - 5 : x(d.barSupport) + 5)
           .attr("y", d => y(d.variable) + y.bandwidth()/2)
           .attr("dy", "0.4em")
-          .attr("text-anchor", d => d.sign == "-1" ? "end" : "start")
+          .attr("text-anchor", d => d.sign == "1" ? "end" : "start")
           .transition()
           .duration(TIME)
           .delay((d,i) => (i+1) * TIME)
-          .text(d => d.sign === "-1" ? d.contribution : "+"+d.contribution);
+          .text(d => d.sign == "-1" ? d.contribution : "+"+d.contribution);
 
   // add lines to bars
   var lines = SV.selectAll()
@@ -508,7 +508,35 @@ function shapValues() {
                              : y(d.variable) + y.bandwidth()*2.5);
 
 
+  // boxplots
+  if (SHOW_BOXPLOT) {
+    // main horizontal line
+    bars.append("line")
+        .attr("class", "interceptLine")
+        .attr("x1", d => d.contribution < 0 ? x(d.max) : x(d.min))
+        .attr("x2", d => d.contribution < 0 ? x(d.max) : x(d.min))
+        .attr("y1", d => y(d.variable) + y.bandwidth()/2)
+        .attr("y2", d => y(d.variable) + y.bandwidth()/2)
+        .transition()
+        .duration(TIME)
+        .delay((d,i) => i * TIME)
+        .attr("x1", d => x(d.min))
+        .attr("x2", d => x(d.max));
 
+    // rectangle for the main box
+    bars.append("rect")
+        .attr("x", d => d.contribution < 0 ? x(d.q3) : x(d.q1))
+        .attr("y", d => y(d.variable) + y.bandwidth()/4)
+        .attr("height", y.bandwidth()/2)
+        .style("fill", "#371ea3")
+        .transition()
+        .duration(TIME)
+        .delay((d,i) => i * TIME)
+        .attr("x", d => x(d.q1))
+        .attr("width", d => Math.abs(x(d.q3) - x(d.q1)));
+  }
+
+  // description
   var description = SV.append("g")
                       .attr("transform", "translate(" +
                             (margin.left + svPlotWidth - 4*margin.big - margin.small)
@@ -604,7 +632,7 @@ function featureImportance() {
             .call(g => g.select(".domain").remove());
 
   var y = d3.scaleBand()
-            .rangeRound([margin.top + fiPlotHeight, margin.top - additionalHeight])
+            .rangeRound([margin.top - additionalHeight, margin.top + fiPlotHeight])
             .padding(0.33)
             .domain(bData.map(d => d.variable));
 
@@ -708,30 +736,33 @@ function featureImportance() {
     .attr("y2", maximumY + y.bandwidth());
 
   // boxplots
-  // main horizontal line
-  bars.append("line")
-      .attr("class", "interceptLine")
-      .attr("x1", d => x(d.dropout_loss) < x(fullModel) ? x(d.max) : x(d.min))
-      .attr("x2", d => x(d.dropout_loss) < x(fullModel) ? x(d.max) : x(d.min))
-      .attr("y1", d => y(d.variable) + y.bandwidth()/2)
-      .attr("y2", d => y(d.variable) + y.bandwidth()/2)
-      .transition()
-      .duration(TIME)
-      .delay((d,i) => i * TIME)
-      .attr("x1", d => x(d.min))
-      .attr("x2", d => x(d.max));
+  if (SHOW_BOXPLOT) {
 
-  // rectangle for the main box
-  bars.append("rect")
-      .attr("x", d => x(d.dropout_loss) < x(fullModel) ? x(d.q3) : x(d.q1))
-      .attr("y", d => y(d.variable) + y.bandwidth()/6)
-      .attr("height", 2*y.bandwidth()/3)
-      .style("fill", "#371ea3")
-      .transition()
-      .duration(TIME)
-      .delay((d,i) => i * TIME)
-      .attr("x", d => x(d.q1))
-      .attr("width", d => Math.abs(x(d.q3) - x(d.q1)));
+    // main horizontal line
+    bars.append("line")
+        .attr("class", "interceptLine")
+        .attr("x1", d => x(d.dropout_loss) < x(fullModel) ? x(d.max) : x(d.min))
+        .attr("x2", d => x(d.dropout_loss) < x(fullModel) ? x(d.max) : x(d.min))
+        .attr("y1", d => y(d.variable) + y.bandwidth()/2)
+        .attr("y2", d => y(d.variable) + y.bandwidth()/2)
+        .transition()
+        .duration(TIME)
+        .delay((d,i) => i * TIME)
+        .attr("x1", d => x(d.min))
+        .attr("x2", d => x(d.max));
+
+    // rectangle for the main box
+    bars.append("rect")
+        .attr("x", d => x(d.dropout_loss) < x(fullModel) ? x(d.q3) : x(d.q1))
+        .attr("y", d => y(d.variable) + y.bandwidth()/4)
+        .attr("height", y.bandwidth()/2)
+        .style("fill", "#371ea3")
+        .transition()
+        .duration(TIME)
+        .delay((d,i) => i * TIME)
+        .attr("x", d => x(d.q1))
+        .attr("width", d => Math.abs(x(d.q3) - x(d.q1)));
+  }
 
   // description
   var description = FI.append("g")
@@ -766,7 +797,7 @@ function featureImportance() {
              .on('mouseout', tooltip.hide);
 }
 
-function partialDependency() {
+function partialDependence() {
 
   if (pdData.x === undefined) return null;
 
@@ -790,7 +821,7 @@ function partialDependency() {
   }
 }
 
-function accumulatedDependency() {
+function accumulatedDependence() {
 
   if (adData.x === undefined) return null;